pyproject.toml
[project]
name = "anchore-security-cli"
requires-python = "<3.14,>=3.13"
license = {text = "Apache 2.0"}
dependencies = [
    "click<9.0.0,>=8.1.3",
    "requests<3.0.0,>=2.28.1",
    "colorlog<7.0.0,>=6.7.0",
    "tomlkit>=0.13.2",
    "python-dateutil>=2.8.2,<3.0.0",
    "pyyaml>=6.0.3",
    "orjson>=3.11.4",
]
dynamic = ["version"]
description = "Tool for performing various Anchore security data curation tasks"
readme = "README.md"
classifiers = [
  # derived from https://pypi.org/classifiers/
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Topic :: Security",
  "Topic :: Utilities",
]

[project.urls]
repository = "https://github.com/anchore/security-cli"

[project.scripts]
anchore-security-cli= "anchore_security_cli.cli.root:root"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "uv-dynamic-versioning"]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"

[tool.uv]
trusted-publishing = "always"

[[tool.uv.index]]
explicit = true
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"

[tool.mypy]
cache_dir = ".cache/mypy"
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
cache_dir = ".cache/pytest"

[tool.ruff.format]
exclude = [
    ".eggs",         # exclude a few common directories in the
    ".git",          # root of the project
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
    "data",
    "backup",
    "src/vunnel/providers/mariner/model", # files in here are generated
    "tests/quality/vulnerability-match-labels",
    "tests/quality/.yardstick",
    "tests/quality/data",
    "tests/quality/build",
    "tests/quality/bin",
]

[tool.ruff]
cache-dir = ".cache/ruff"
line-length = 150
extend-exclude = [
  "**/tests/**",
  "src/vunnel/providers/mariner/model/**" # these are generated
]

[tool.ruff.lint]
ignore = [
  "ARG001", # unused args are ok, as they communicate intent in interfaces, even if not used in impls.
  "ARG002", # unused args are ok, as they communicate intent in interfaces, even if not used in impls.
  "G004", # it's ok to use formatted strings for logging
  "PGH004", # no blanket "noqa" usage, can be improved over time, but not now
  "PLR2004", # a little too agressive, not allowing any magic numbers
  "PLW2901", # "Outer for loop variable X overwritten by inner assignment target", not useful in most cases
  "RUF100", # no blanket "noqa" usage, can be improved over time, but not now
  "S603", # subprocess calls are common and necessary in this codebase
  "S607", # we cannot rely on exact paths to all tools invoked for all users so must rely on PATH resolution
]

select = [
  "A",   # flake8-builtins
  # "ANN", # flake8-annotations   # this is great, but let mypy handle this so it can honor type:ignore comments without noqa comments too
  "ARG", # flake8-unused-arguments
  "B",   # flake8-bugbear
  "C",   # mccabe
  "C4",  # flake8-comprehensions
  "COM", # flake8-commas
  "DTZ", # flake8-datetimez
  "E",   # pycodestyle, errors
  # "EM",  # flake8-errmsg     # seems to aggressive relative to the current codebase
  # "ERA", # flake8-eradicate  # not all code in comments should be removed
  "EXE", # flake8-executable
  "F",   # pyflakes
  "G",   # flake8-logging-format
  "I",   # isort
  "ICN", # flake8-import-conventions
  "INP", # flake8-no-pep420
  "ISC", # flake8-implicit-str-concat
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  "PL",  # pylint     (this can be broken down into more checks if needed)
  "PT",  # flake8-pytest-style
  # "PTH", # flake8-use-pathlib  # the codebase is not ready for this yet, but would be nice to add
  "Q",   # flake8-quotes
  "RET", # flake8-return
  "RSE", # flake8-raise
  "RUF", # ruff specific rules
  "S",   # flake8-bandit
  "SIM", # flake8-simplify
  "SLF", # flakes8-self
  "T10", # flake8-debugger
  "TCH", # flake8-type-checking
  "TID", # flake8-tidy-imports
  # "TRY", # tryceratops   # seems to aggressive relative to the current codebase
  "UP",  # pyupgrade
  "W",   # pycodestyle, warnings
  "YTT", # flake8-2020
]

[dependency-groups]
dev = [
    "pytest-mock<4.0.0,>=3.14.0",
    "pytest<9.0.0,>=8.3.4",
    "pre-commit<5.0.0,>=4.0.1",
    "mypy<2.0,>=1.13",
    "isort<6.0.0,>=5.13.2",
    "pylint<4.0.0,>=3.3.2",
    "autoflake<3.0,>=2.3",
    "ruff>=0.8.1,<1.0.0",
    "pytest-unordered<1.0.0,>=0.6.1",
    "pytest-sugar<2.0.0,>=1.0.0",
    "pytest-clarity<2.0.0,>=1.0.1",
    "pytest-cov<7.0.0,>=6.0.0",
    "pytest-picked<1.0.0,>=0.5.0",
    "pytest-xdist<4.0.0,>=3.6.1",
    "types-requests<3.0.0.0,>=2.32.0.20241016",
    "types-tabulate<1.0.0.0,>=0.9.0.20240106",
    "types-pyyaml<7.0.0.0,>=6.0.12.20240917",
    "types-pygments<3.0.0.0,>=2.18.0.20240506",
    "hatchling>=1.27.0",
    "hatch-vcs>=0.4.0",
]
